+++ voting/frontend/app/admin/page.tsx
@@ -1,88 +1,45 @@
-"use client";
+"use client"
 
-import { useAccount } from "wagmi"
-import { useEffect, useState } from "react"
-import { readContractByFunctionName } from "@/utils"
-import { useRouter } from 'next/navigation';
-import IsConnected from "@/components/IsConnected";
-import WorkflowManager from "@/components/admin/WorkflowManager";
+import IsConnected from "@/components/IsConnected"
+import WinningProposal from "@/components/WinningProposal"
+import VoterManager from "@/components/admin/VoterManager"
+import WorkflowManager from "@/components/admin/WorkflowManager"
+
+import { WorkflowStatus, registeringVotersStatus, votesTalliedStatus } from "@/constants"
+import { useWorkflowStatusContext } from "@/context/workflowStatus"
+import { getWorkflowStatus } from "@/utils"
+import { useEffect } from "react"
+import {useAccount} from "wagmi";
+import Event from "@/components/Event";
+import {WorkflowStepperManager} from "@/components/home/WorkflowStepperManager";
 
 const Admin = () => {
-  const { address, isConnected } = useAccount()
-  const { push } = useRouter()
-  const [owner, setOwner] = useState('')
-  const [winningProposalID, setWinningProposalID] = useState<number|null>(null)
-  const [isOwner, setIsOwner] = useState(false)
-
-  const getOwner = async () => {
-    readContractByFunctionName<`0x${string}`>('owner', address as `0x${string}`).then(
-      hash => {
-        if (hash !== address) push('/')
-        setOwner(hash)
-        setIsOwner(hash === address)
-    }).catch(() => push('/'))
-  }
-
-  const getWinningProposalID = async () => {
-    readContractByFunctionName<number>('winningProposalID', address as `0x${string}`).then(
-        id => setWinningProposalID(id)
-    ).catch(
-        err => console.log(err.message)
-    )
-  }
+  const { address } = useAccount()
+  const { workflowStatus, setWorkflowStatus } = useWorkflowStatusContext()
 
   useEffect(() => {
-    if (isConnected) {
-      getOwner()
-      getWinningProposalID()
-    }
-  }, [address, isConnected])
+    getWorkflowStatus(address as `0x${string}`).then(
+      id => setWorkflowStatus(id)
+    ).catch(err => console.log(err))
+  }, [workflowStatus])
 
   return (
-    <div className="flex flex-col space-y-2 mx-auto max-w-screen-lg">
-      <IsConnected>
-        {isOwner ? (
-          <>
-            <div className="mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-600 text-indigo-100 shadow-lg">
-              <div className="p-4">
-                <ul className="list-inside list-disc">
-                  <li>{owner} (Contract's owner)</li>
-                  <li>{address} (Account address)</li>
-                </ul>
+    <IsConnected asOwner={true}>
+        <WorkflowStepperManager />
+        <WorkflowManager />
 
-              </div>
-            </div>
+        <div className="mx-auto w-3/4">
+            {WorkflowStatus[workflowStatus] === votesTalliedStatus && <WinningProposal />}
+        </div>
 
-            <div className="mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-600 text-indigo-100 shadow-lg">
-              <div className="p-4">
-                winningProposalID : {winningProposalID}
-              </div>
-            </div>
+        {WorkflowStatus[workflowStatus] === registeringVotersStatus && <VoterManager />}
 
-            <WorkflowManager />
-          </>
+        <div className="mt-2">
+            <Event name='VoterRegistered'></Event>
+            <Event name='WorkflowStatusChange'></Event>
+        </div>
 
-        ) : (
-          <div className="mx-auto w-3/4 rounded h-[90px] bg-gradient-to-r from-indigo-950 to-rose-600 text-zinc-200 shadow-lg">
-            <div className="flex justify-between pt-2">
-              <div className="px-4 font-light text-sm">
-                You are not the owner !
-              </div>
-            </div>
-            <div className="flex justify-between pt-2">
-              <div className="px-4 font-light text-sm">
-                The owner is {owner}
-              </div>
-            </div>
-            <div className="flex justify-between pt-2">
-              <div className="px-4 font-light text-sm">
-                You are {address}
-              </div>
-            </div>
-          </div>
-        )}
-      </IsConnected>
-    </div>
+    </IsConnected>
   )
 }