+++ voting/frontend/app/voters/page.tsx
@@ -1,33 +1,25 @@
 "use client";
 
-import Event from "@/components/Event";
 import IsConnected from "@/components/IsConnected";
 import GetVoter from "@/components/voters/GetVoter";
 import ProposalManager from "@/components/voters/ProposalManager";
-import { userIsVoter } from "@/utils";
-import { useRouter } from 'next/navigation';
-import { useEffect } from "react";
-import { useAccount } from "wagmi";
-import ProposalEvents from "@/components/voters/ProposalEvents";
+import SendVote from "@/components/voters/SendVote";
+import Event from "@/components/Event";
+import {WorkflowStepperManager} from "@/components/home/WorkflowStepperManager";
+
 
 export default function Voters() {
-  const { address, isConnected } = useAccount()
-  const { push } = useRouter()
-
-  useEffect(() => {
-    if (isConnected) userIsVoter(address as `0x${string}`).then(
-      isVoter => {if (!isVoter) push('/')}
-    ).catch(() => push('/'))
-  }, [address, isConnected])
-
   return (
-      <>
-        <IsConnected>
-            <GetVoter />
-            <ProposalManager />
-            <ProposalEvents/>
+
+    <IsConnected asVoter={true}>
+        <WorkflowStepperManager />
+        <GetVoter />
+        <ProposalManager />
+        {/*<SendVote/>*/}
+
+        <div className="mt-2">
             <Event name='ProposalRegistered'></Event>
-        </IsConnected>
-      </>
+        </div>
+    </IsConnected>
   )
 }