+++ voting/frontend/components/admin/WorkflowManager.tsx
@@ -1,41 +1,59 @@
 "use client"
 
-import {useEffect, useState} from "react"
-import {readContractByFunctionName, writeContractByFunctionName} from "@/utils";
-import {useWorkflowStatusContext} from "@/context/workflowStatus";
-import VoterManager from "@/components/admin/VoterManager";
-import Event from "@/components/Event";
-import {useToast} from "@chakra-ui/react";
+import { useEffect, useState } from "react"
+import { getWorkflowStatus, writeContractByFunctionName } from "@/utils";
+import { useWorkflowStatusContext } from "@/context/workflowStatus";
+import { useToast } from "@chakra-ui/react";
+import { WorkflowStatus, abi, contractAddress, proposalsRegistrationEndedStatus, proposalsRegistrationStartedStatus, registeringVotersStatus, votesTalliedStatus, votingSessionEndedStatus, votingSessionStartedStatus } from "@/constants";
+import { useContractEvent } from "wagmi";
+import { Log } from "viem";
+import Loader from "../Loader";
 import {useAccount} from "wagmi";
 
 const WorkflowManager = () => {
-    const { address, isConnected } = useAccount()
-    const [success, setSuccess] = useState('')
-    const { workflowStatus, setWorkflowStatus} = useWorkflowStatusContext()
+    const [logs, setLogs] = useState<Log[]>()
+    const [loading, setLoading] = useState(true)
+    const { address } = useAccount()
+    const { workflowStatus, setWorkflowStatus } = useWorkflowStatusContext()
     const toast = useToast()
 
-    const WorkflowStatus: string[] = [
-        "RegisteringVoters",
-        "ProposalsRegistrationStarted",
-        "ProposalsRegistrationEnded",
-        "VotingSessionStarted",
-        "VotingSessionEnded",
-        "VotesTallied"
-    ]
+    useContractEvent({
+        address: contractAddress,
+        abi: abi,
+        eventName: 'WorkflowStatusChange',
+        listener(log) {
+            setLogs(log)
+        }
+    })
 
     useEffect(() => {
-        getWorkflowStatus()
-    }, [success])
+        getWorkflowStatus(address as `0x${string}`).then(
+            id => setWorkflowStatus(id)
+        ).catch(err => console.log(err))
+        .finally(() => setLoading(false))
+    }, [logs])
 
-    const updateWorkflowStatus = (funcName: string) => {
-        setSuccess('')
+    const nextWorkflowStatus = () => {
+        let funcName = '';
+        switch (WorkflowStatus[workflowStatus]) {
+            case registeringVotersStatus: funcName = 'startProposalsRegistering'
+                break
+            case proposalsRegistrationStartedStatus: funcName = 'endProposalsRegistering'
+                break
+            case proposalsRegistrationEndedStatus: funcName = 'startVotingSession'
+                break
+            case votingSessionStartedStatus: funcName = 'endVotingSession'
+                break
+            case votingSessionEndedStatus: funcName = 'tallyVotes'
+                break
+        }
 
         writeContractByFunctionName(funcName).then(
             hash => {
-                setSuccess(hash)
+                setLoading(true)
                 toast({
                     title: 'Workflow status updated.',
-                    description: `with status: ${funcName}.`,
+                    description: `Transaction: ${hash}`,
                     status: 'success',
                     duration: 5000,
                     isClosable: true,
@@ -52,70 +70,23 @@ const WorkflowManager = () => {
         )
     }
 
-    const getWorkflowStatus = async () => {
-        readContractByFunctionName<number>('workflowStatus', address as `0x${string}`).then(
-            id => setWorkflowStatus(id)
-        ).catch(
-            err => console.log(err.message)
-        )
-    }
-
     return (
-        <>
-            <div className="mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-800 text-indigo-100 shadow-lg drop-shadow-lg border-indigo-600 border">
+        <Loader isLoading={loading}>
+            <div className="mt-2 mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-800 text-indigo-100 shadow-lg drop-shadow-lg border-indigo-600 border">
                 <div className="p-4 font-bold text-md">
-                    WorkflowStatus : {WorkflowStatus[workflowStatus]}
+                    Workflow Status : {WorkflowStatus[workflowStatus]}
                 </div>
             </div>
 
-            {WorkflowStatus[workflowStatus] === "RegisteringVoters" && <VoterManager /> }
-
-            {WorkflowStatus[workflowStatus] !== "VotesTallied" &&
-
-                <div className="mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-800 text-indigo-100 shadow-lg drop-shadow-lg border-indigo-600 border">
-                    <div className="p-4">
-
-                        {WorkflowStatus[workflowStatus] === "RegisteringVoters" &&
-                            <button
-                                onClick={() => updateWorkflowStatus("startProposalsRegistering")}
-                                className="bg-indigo-950 hover:bg-indigo-100 hover:text-gray-900 text-white font-semibold py-2 px-4 rounded-lg"
-                            >Start Proposals Registering</button>
-                        }
-
-                        {WorkflowStatus[workflowStatus] === "ProposalsRegistrationStarted" &&
-                            <button
-                                onClick={() => updateWorkflowStatus("endProposalsRegistering")}
-                                className="bg-indigo-950 hover:bg-indigo-100 hover:text-gray-900 text-white font-semibold py-2 px-4 rounded-lg"
-                            >End Proposals Registering</button>
-                        }
-
-                        {WorkflowStatus[workflowStatus] === "ProposalsRegistrationEnded" &&
-                            <button
-                                onClick={() => updateWorkflowStatus("startVotingSession")}
-                                className="bg-indigo-950 hover:bg-indigo-100 hover:text-gray-900 text-white font-semibold py-2 px-4 rounded-lg"
-                            >Start voting session</button>
-                        }
-
-                        {WorkflowStatus[workflowStatus] === "VotingSessionStarted" &&
-                            <button
-                                onClick={() => updateWorkflowStatus("endVotingSession")}
-                                className="bg-indigo-950 hover:bg-indigo-100 hover:text-gray-900 text-white font-semibold py-2 px-4 rounded-lg"
-                            >End voting session</button>
-                        }
-
-                        {WorkflowStatus[workflowStatus] === "VotingSessionEnded" &&
-                            <button
-                                onClick={() => updateWorkflowStatus("tallyVotes")}
-                                className="bg-indigo-950 hover:bg-indigo-100 hover:text-gray-900 text-white font-semibold py-2 px-4 rounded-lg"
-                            >Tally votes</button>
-                        }
-                    </div>
-
+            {WorkflowStatus[workflowStatus] !== votesTalliedStatus &&
+                <div className="mt-2 p-4 mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-800 text-indigo-100 shadow-lg drop-shadow-lg border-indigo-600 border ">
+                    <button onClick={() => nextWorkflowStatus()} className="bg-indigo-950 hover:bg-indigo-100 hover:text-gray-900 text-white font-semibold py-2 px-4 rounded-lg">
+                        Next Step: {WorkflowStatus[workflowStatus + 1]}
+                    </button>
                 </div>
             }
 
-            <Event name='WorkflowStatusChange'></Event>
-        </>
+        </Loader>
     )
 }
 export default WorkflowManager