+++ voting/frontend/interfaces/Proposal.ts
@@ -1,0 +1,4 @@
+export interface Proposal {
+    description: string
+    voteCount: bigint
+}